Cleaning up threading in HttpClient and PollSet
Introduced usage of thread safety analysis from CLang. It helped
tracking some of potential issues in HttpClient, but created a lot of
side problems with usage of other libc++.

Fixed typo with setting MINIROS_COMPILE_OPTIONS. Now required level of
warnings is properly set.

Moved all small helpers related to threading into internal/threading.h
header. Not it also defines all markup macros required by thread safety
analysis from CLang.

Added flag to PollSet to start without internal signaling.

Fixed some of compiler warnings.

+++ include/miniros/names.h
@@ -122,10 +122,10 @@ struct MINIROS_DECL Path {
   std::string name() const;
 
   /// Get a name from position [i, end]
-  std::string right(int i) const;
+  std::string right(size_t i) const;
 
   /// Get a path [0, i)
-  std::string left(int i) const;
+  std::string left(size_t i) const;
 
   /// Returns a string with a full original path.
   const std::string& fullPath() const;
